+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 - Add new `predbox` base. These bases are labels for the [Parametric Gridfinity Storage Box by Pred][predbox]
   box labels. It is supported for width 4, 5, 6 and 7, which corresponds to the
   label size corresponding to a storage box of that many gridfinity units.
+- Add fragment `{|}`. This allows you to designate columns between which
+  the text area will be split.
 
 [predbox]: https://www.printables.com/model/543553-gridfinity-storage-box-by-pred-now-parametric